pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-usage-widget"
dynamic = ["version"]
description = "Desktop widget and CLI that shows real-time Claude Code usage limits and cost."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Burak" }]
keywords = ["claude", "anthropic", "usage", "rate-limit", "widget"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Environment :: MacOS X",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]

# Single runtime dependency: PySide6-Essentials bundles Qt (no apt/brew
# needed).  Notifications use the platform's built-in CLI (notify-send on
# Linux, osascript on macOS), so no extra deps for that either.
dependencies = [
    "PySide6-Essentials>=6.5,<7",
]

[project.urls]
Homepage = "https://github.com/bozdemir/claude-usage-widget"
Issues = "https://github.com/bozdemir/claude-usage-widget/issues"

[project.scripts]
claude-usage = "claude_usage.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["claude_usage*"]

[tool.setuptools.dynamic]
version = { attr = "claude_usage.__version__" }

[tool.setuptools.package-data]
claude_usage = ["py.typed", "icons/*.svg"]
